+++ packages/cli/src/platform/$.ts
@@ -15,7 +15,7 @@ export class $ implements Command {
       ['auth', 'Manage authentication with your Prisma Data Platform account'],
       ['workspace', 'Manage workspaces'],
       ['project', 'Manage projects'],
-      ['apikey', 'Manage API keys'],
+      ['serviceToken', 'Manage service tokens'],
       ['accelerate', 'Manage Prisma Accelerate'],
     ],
     options: [
