+++ packages/cli/src/platform/*.ts
@@ -1,8 +1,8 @@
 export * from './$'
 export * from './accelerate'
-export * from './apikey/*APIKey'
 export * from './auth/*Auth'
 export * from './auth/login'
 export * from './auth/logout'
 export * from './project/*Project'
+export * from './serviceToken/*ServiceToken'
 export * from './workspace/*Workspace'
