--- packages/cli/src/platform/apikey/*APIKey.ts
@@ -1,1 +1,0 @@
-export * as APIKey from './*'
