+++ packages/cli/src/platform/serviceToken/$.ts
