+++ packages/cli/src/platform/serviceToken/*ServiceToken.ts
@@ -1,0 +1,1 @@
+export * as ServiceToken from './*'
