+++ packages/cli/src/platform/serviceToken/create.ts
