+++ packages/cli/src/platform/serviceToken/show.ts
